requirements.txt
Django>=4.0.1,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.26.1,<0.27
Pillow>=9.1.0,<9.2
uwsgi>=2.0.20,<2.1
beautifulsoup4>=4.10.0,<4.11
requests>=2.26.0,<2.27
django-cors-headers>=3.8.0,<3.9
django-filter>=2.4.0,<2.5
djangorestframework-simplejwt>=5.2.0,<5.3
django-allauth>=0.52.0,<0.53
dj-rest-auth==4.0.0
django-rest-passwordreset>=1.3.0,<1.4